Relax the `Sized` constraint in `impl_refs`
The traits implemented using the `impl_refs` were accidentally bound by
`Sized` which shouldn't be required since these are the reference types.
This change removes the restriction. This makes the macro no longer work
for traits that do require `Sized` but none are actually used so far and
a new macro can solve it in the future.