Document when using `NonEmpty<&mut T>` makes sense
The documentation so far stated that `&mut NonEmpty<T>` is preferred to
`NonEmpty<&mut T>`, however this doesn't work when the type state needs
to change during function execution. In effect, `&mut NonEmpty<T>`
states that the contained `T` will stay non-empty for the lifetime of
the reference while `NonEmpty<&mut T>` does not.

To avoid confusion, this is now documented.